Rename `workouts-table-dialog-types.ts` file to `types.ts` and update imports
